% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eap_region}
\alias{eap_region}
\alias{eap_countries}
\title{Study region and countries}
\format{An \code{sf} dataset}
\usage{
data(eap_region)
}
\description{
These points represent population-weighted centroids of Medium Super Output Area (MSOA) zones within a 1 mile radius of of my home when I was writing this package.
}
\details{
Cents was generated from the data repository pct-data: https://github.com/npct/pct-data. This data was accessed from within the pct repo: https://github.com/npct/pct, using the following code:
}
\examples{
\dontrun{
library(sf)
library(tidyverse)
eap_countries = c("Armenia", "Azerbaijan", "Belarus", "Georgia", "Moldova", "Ukraine")
eap_countries = spData::world \%>\%
  filter(name_long \%in\% eap_countries)
mapview::mapview(eap_countries)
devtools::use_data(eap_countries)
eap_buff = st_buffer(eap_countries, 0.00001)
eap_region = st_difference(eap_buff, eap_buff) \%>\% 
  st_combine() \%>\% 
  st_union()
plot(eap_region)
devtools::use_data(eap_region)
}
}
\keyword{datasets}
